# swagger.yaml
"openapi": "3.0.3"
"info":
  "title": "team-work hub"
  "description": "API for team work hub project."
  
  "contact":
    "email": "mr.oshanbiswas@gmail.com"
  
  "version": "1.0.11"
"servers":
  -
    "url": "http://localhost:4000/api/v1"
    "description": "Local server"
...
"paths":
...
  "/content/feedback/{id}":
    "post":
      "summary": "Add feedback to a document"
      "tags":
        - "Feedback"
      "security":
        -
          "bearerAuth": []
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "required":
                - "user"
                - "feedbackText"
                - "content"
              "properties":
                "user":
                  "type": "string"
                  "example": "bhjsheueh3412j231231n"
                "feedbackText":
                  "type": "string"
                "content":
                  "type": "string"
                  "example": "12m1j2k1j121j3jjjj3"
      "responses":
        "201":
          "description": "Feedback added successfully"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "acknowlwdged":
                    "type": "boolean"
                  "insertedId":
                    "type": "string"
                    "example": "hdb1b3j1b12jb34j4j4"
        "400":
          "description": "Bad Request"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "errors":
                    "type": "array"
                    "items":
                      "type": "object"
                      "properties":
                        "code":
                          "type": "string"
                          "example": "invalid_type"
                        "expected":
                          "type": "string"
                          "example": "string"
                        "received":
                          "type": "string"
                          "example": "undefined"
                        "path":
                          "type": "array"
                          "items":
                            "type": "string"
                            "example": "user"
                        "message":
                          "type": "string"
                          "example": "Required"
        "401":
          "description": "Unauthorized"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "Unauthorized"
        "500":
          "description": "Internal server error"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "Internal server error"
    "get":
      "summary": "Get details of a feedback"
      "tags":
        - "Feedback"
      "security":
        -
          "bearerAuth": []
      "description": "Retrieve details of a feedback. This route is protected and requires authorization."
      "operationId": "getFeedbackDetails"
      "parameters":
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
      "responses":
        "200":
          "description": "Feedback data with details of the feedback"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "_id":
                    "type": "string"
                    "example": "12m1j2k1j121j3jjjj3"
                  "user":
                    "type": "string"
                    "example": "bhjsheueh3412j231231n"
                  "body":
                    "type": "string"
                  "content":
                    "type": "string"
                    "example": "12m1j2k1j121j3jjjj3"
                  "createdAt":
                    "type": "string"
                    "format": "date"
                    "example": "2024-06-28"
                  "updatedAt":
                    "type": "string"
                    "format": "date"
                    "example": "2024-06-28"
        "404":
          "description": "not found"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "not found"
        "500":
          "description": "Internal server error"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "Internal server error"
    "patch":
      "summary": "Update a feedback"
      "tags":
        - "Feedback"
      "security":
        -
          "bearerAuth": []
      "parameters":
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "properties":
                "feedbackText":
                  "type": "string"
      "responses":
        "200":
          "description": "Feedback updated successfully"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "acknowlwdged":
                    "type": "boolean"
                  "updatedId":
                    "type": "string"
                    "example": "hdb1b3j1b12jb34j4j4"
                  "matchedCount":
                    "type": "number"
                    "example": 1
                  "updatedCount":
                    "type": "number"
                    "example": 1
        "400":
          "description": "Bad Request"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "errors":
                    "type": "array"
                    "items":
                      "type": "object"
                      "properties":
                        "code":
                          "type": "string"
                          "example": "invalid_type"
                        "expected":
                          "type": "string"
                          "example": "string"
                        "received":
                          "type": "string"
                          "example": "undefined"
                        "path":
                          "type": "array"
                          "items":
                            "type": "string"
                            "example": "body"
                        "message":
                          "type": "string"
                          "example": "Required"
        "401":
          "description": "Unauthorized"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "Unauthorized"
        "500":
          "description": "Internal server error"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "Internal server error"
        "404":
          "description": "not found"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "message":
                    "type": "string"
                    "example": "not found"
    "delete":
      "summary": "Delete a feedback"
      "tags":
        - "Feedback"
      "security":
        -
          "bearerAuth": []
      "description": "Delete a Feedback by ID. This route is protected."
      "operationId": "deleteFeedback"
      "parameters":
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
      "responses":
        "204":
          "description": "Content data with details of the content"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "properties":
                  "_id":
                    "type": "string"
                    "example": "12m1j2k1j121j3jjjj3"
                  "acknowledged":
                    "type": "boolean"
                  "deletedCount":
                    "type": "number"
                    "example": 1
        "404":
          "description": "Feedback not found"
        "401":
          "description": "Unauthorized"
"components":
  "securitySchemes":
    "bearerAuth":
      "type": "http"
      "scheme": "bearer"
...